change: switch the default Envoy version to 1.32
