+++ api/pkg/filtermanager/api_impl_129.go
@@ -33,6 +33,10 @@ func (cb *filterManagerCallbackHandler) ClearRouteCache() {
 	api.LogErrorf("ClearRouteCache is not implemented: %s", debug.Stack())
 }
 
+func (cb *filterManagerCallbackHandler) RefreshRouteCache() {
+	api.LogErrorf("RefreshRouteCache is not implemented: %s", debug.Stack())
+}
+
 func (cb *filterManagerCallbackHandler) DecoderFilterCallbacks() api.DecoderFilterCallbacks {
 	return cb.FilterCallbackHandler
 }
