+++ api/pkg/filtermanager/filtermanager_latest.go
@@ -12,53 +12,35 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build !envoydev
+//go:build !envoy1.29 && !envoy1.31
 
 package filtermanager
 
 import (
 	capi "github.com/envoyproxy/envoy/contrib/golang/common/go/api"
-
-	"mosn.io/htnn/api/pkg/filtermanager/api"
 )
 
 const (
-	supportGettingHeadersOnLog = false
+	supportGettingHeadersOnLog = true
 )
 
-func (m *filterManager) OnLog(_ capi.RequestHeaderMap, _ capi.RequestTrailerMap, _ capi.ResponseHeaderMap, _ capi.ResponseTrailerMap) {
+func (m *filterManager) OnLog(reqHdr capi.RequestHeaderMap, reqTrailer capi.RequestTrailerMap, rspHdr capi.ResponseHeaderMap, rspTrailer capi.ResponseTrailerMap) {
 	if m.canSkipOnLog {
 		return
 	}
 
-	var reqHdr api.RequestHeaderMap
-	m.hdrLock.Lock()
-	reqHdr = m.reqHdr
-	m.hdrLock.Unlock()
-	var rspHdr api.ResponseHeaderMap
-	m.hdrLock.Lock()
-	rspHdr = m.rspHdr
-	m.hdrLock.Unlock()
-
-	m.runOnLogPhase(reqHdr, nil, rspHdr, nil)
+	wrappedReqHdr := &filterManagerRequestHeaderMap{
+		RequestHeaderMap: reqHdr,
+	}
+	m.runOnLogPhase(wrappedReqHdr, reqTrailer, rspHdr, rspTrailer)
 }
-
-type filterManagerWrapper struct {
-	*filterManager
-}
-
-func (w *filterManagerWrapper) OnLog() {
-	w.filterManager.OnLog(nil, nil, nil, nil)
-}
-
-// we will get rid of this wrapper once Envoy 1.32 is released
 
 func wrapFilterManager(fm *filterManager) capi.StreamFilter {
-	return &filterManagerWrapper{fm}
+	return fm
 }
 
-// This method is test only
+// This method is only for test
 func unwrapFilterManager(wrapper capi.StreamFilter) *filterManager {
-	fmw, _ := wrapper.(*filterManagerWrapper)
-	return fmw.filterManager
+	p, _ := wrapper.(*filterManager)
+	return p
 }
