+++ common.mk
@@ -29,8 +29,8 @@ DOCKER_MIRROR   = m.daocloud.io/
 
 # Both images use glibc 2.31. Ensure libc in the images match each other.
 BUILD_IMAGE     ?= $(DOCKER_MIRROR)docker.io/library/golang:1.22-bullseye
-ENVOY_API_VERSION ?= 1.31
-PROXY_IMAGE     ?= $(DOCKER_MIRROR)docker.io/envoyproxy/envoy:contrib-v1.31.2
+ENVOY_API_VERSION ?= 1.32
+PROXY_IMAGE     ?= $(DOCKER_MIRROR)docker.io/envoyproxy/envoy:contrib-v1.32.0
 # We also support other Envoy versions. See https://github.com/mosn/htnn/tree/main/site/content/en/docs/developer-guide/dataplane_support.md
 
 # We may need to use timestamp if we need to update the image in one PR
