+++ controller/internal/translation/data_plane_state_test.go
@@ -162,7 +162,6 @@ func TestAllowRoute(t *testing.T) {
 	}
 
 	for _, tt := range tests {
-		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			actual := AllowRoute(tt.cond, route, gwNsName)
 			if actual != tt.expected {
