+++ e2e/go.mod
@@ -14,7 +14,7 @@
 
 module mosn.io/htnn/e2e
 
-go 1.21.5
+go 1.22
 
 require (
 	github.com/avast/retry-go v3.0.0+incompatible
